[#7231] Add support for multi flow exporter targets via CRDs.
This change uses the new CRD `FlowExporterTarget`. The major changes
introduced in this commit is the merging of the two connection stores
(CT and deny connections). The store is now only responsible to keep
track of the latest state while the consumers (the exporters) are
responsible for tracking what additional dynamic data it needs to
populate.

There are some additional cleanups and other follow up:
- Removing unused fields in the `connection` struct.
- Add migration documentations
- Remove unused configs from Antrea ConfigMap