Increase the minimum supported Kubernetes version to 1.23 (#7564)
K8s 1.23 is the release where the IPv6DualStack feature gate became GA
and locked to true. By increasing the minimum supported version to 1.23,
we can assume that some dual-stack fields (e.g., spec.ipFamilies` for
Services) are always present and populated correctly.

K8s 1.23 is also the release were the EndpointSlice feature gate became
GA and locked to true.

Additionally, because NamespaceDefaultLabelName was removed in K8s 1.23,
we can furthermore remove some code / documentation that is no longer
needed. However, we do not drop support for `antrea.io/metadata.name`
(yet?) - we just stop mentioning it in the Antrea NetworkPolicy
documentation.

Note that K8s 1.23 was released at the end of 2021.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>