Fix conformance failures for netpolv2 (#7563) (#7581)
Due to broken dependencies in upstream, the netpol
v2 conformance tests were always executed against
code from a fork. Recently the main branch of the
fork was synced with the upstream main, causing
conformance tests to break since the upstream main
branch is now only testing conformance for a new,
un-released ClusterNetworkPolicy API type.

Fixing the failure by running the conformance test
against the API from the correct release tag. When
support for ClusterNetworkPolicy in Antrea is
implemented, the script will be updated to only
test conformance for the new resource instead.

Signed-off-by: Dyanngg <dingyang@vmware.com>