Re-enable NodePortLocal e2e tests (#7589)
Ever since #5491 was merged, it seems that NPL e2e tests have never
run. The test cases were being skipped because nodePortLocal.enable was
always false in the Agent configuration. The logic meant to set
nodePortLocal.enable to true was still present in the test code, but
after the check that was causing the tests to be skipped.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>