Fix panic in `antctl check cluster` (#7595)
The panic was introduced by #7564.
Prior to that, `semver.Parse` was used instead of `semver.MustParse`,
and the returned error was ignored. While there was no panic prior to
the change, the code was still incorrect as the K8s server version was
always compared to 0.0.0, instead of the actual min required version.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>