Fix "antctl check installation" in CI
The CI job was consistently failing because the check command was run
right after installing Antrea, without waiting for the Antrea components
to become ready. In particular, at least some of the Agents didn't have
a succesful watch to the Controller (and there is a 5s backoff before
retrying), preventing new NetworkPolicies from being distributed
successfully.

To avoid the issue, we run "kubectl rollout status" for the
antrea-controller Deployment and for the antrea-agent DaemonSet, after
installing Antrea and before running the check command. For good
measure, we also update the command implementation to check (and wait)
for availability of the Antrea components before running validation
tests.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>