Fix Traceflow with WireGuard enabled
In hybrid mode, packets are forwarded directly. In WireGuard mode, traffic
is encapsulated by WireGuard. In both modes, traffic leaves OVS, but
Traceflow was not checking for WireGuard mode when determining packet actions.

This commit adds WireGuard mode checks in Traceflow to correctly handle
packets when WireGuard encryption is enabled.

Signed-off-by: xliuxu <xu.liu@broadcom.com>