Stop exporting AppProtocolName and HttpVals
We deprecate the App field in the Flow Protobuf used to export records
from the Agent to the Flow Aggregator. We stop exporting App data in the
Agent (Flow Exporter), and we stop consuming App data in the Flow
Aggregator. In the Flow Aggregator, the various exporters no longer
export the corresponding fields. The IPFIX exporter no longer includes
the corresponding Information Elements. The ClickHouse client no longer
sets the corresponding fields in insert statements. The S3 uploader no
longer includes the corresponding fields in CSV records. Note that Theia
was never updated to support L7 fields (appProtocolName and
httpVals). Other potential consumers of the S3 uploader may need to be
updated after this change (not backward-compatible), which we think is
acceptable.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>