[FlowAggregator] Generate self-signed certificates only if needed (#7559)
This addresses an issue where having multiple replicas of
flow-aggregator will cause it to clobber each other's
certificates. We now only try to create it if it is invalid
(does not exist/expired). The root CA certificate and key
is shared amongst all instances. Each replica is responsible
for generating their own server certificates. Client certs
and the CA ConfigMap will be created/updated if it is invalid
or is incompatible with the CA certificate.

e2e tests will use multiple replicas when testing the
flow-aggregator in Proxy mode.

Fixes #7540