Fix testReconcileGatewayRoutesOnStartup with WireGuard (#7635)
The test was failing when WireGuard encryption is enabled because it
expects gateway routes for encap mode, but WireGuard doesn't require
gateway routes since it handles routing through its own interface.

This commit adds encryption mode detection to the test framework and
updates the test to expect 0 gateway routes when WireGuard is enabled
with encap mode.

Signed-off-by: xliuxu <xu.liu@broadcom.com>